Clean up HomeKit SonarQube findings
- use enum class for the HomeKit mode setting
- move HomeKit settings migration out of the Settings constructor
- split HomeKit bridge setup into smaller helper methods
- replace Bridge-specific macros and typedefs with constexpr/using
- improve HomeKit settings accessibility and parsing style
- keep existing HomeSpan and plugin allocation patterns unchanged

+++ src/display/plugins/WebUIPlugin.cpp
@@ -607,7 +607,7 @@ void WebUIPlugin::handleSettings(AsyncWebServerRequest *request) const {
     doc["startupMode"] = settings.getStartupMode() == MODE_BREW ? "brew" : "standby";
     doc["targetSteamTemp"] = settings.getTargetSteamTemp();
     doc["targetWaterTemp"] = settings.getTargetWaterTemp();
-    doc["homekitMode"] = settings.getHomekitMode();
+    doc["homekitMode"] = static_cast<int>(settings.getHomekitMode());
     doc["hkPowerEnabled"] = settings.isHkPowerEnabled();
     doc["hkSteamEnabled"] = settings.isHkSteamEnabled();
     doc["hkSensorEnabled"] = settings.isHkSensorEnabled();
